Move Reserve and MsgMapWidget to new pub kas::adapter module
