+++ src/widget/mod.rs
@@ -66,7 +66,6 @@ mod menu;
 mod nav_frame;
 mod progress;
 mod radiobox;
-mod reserve;
 mod scroll;
 mod scrollbar;
 mod separator;
@@ -91,7 +90,6 @@ pub use menu::*;
 pub use nav_frame::NavFrame;
 pub use progress::ProgressBar;
 pub use radiobox::{RadioBox, RadioBoxBare};
-pub use reserve::{Reserve, ReserveP};
 pub use scroll::{ScrollComponent, ScrollRegion};
 pub use scrollbar::{ScrollBar, ScrollBarRegion, ScrollBars, Scrollable};
 pub use separator::Separator;
